+++ src/app/(public)/page.tsx
@@ -1,5 +1,12 @@
+import { JsonLd } from 'features/seo/ui/JsonLD';
 import { HomePage } from 'pages/public/HomePage';
+import { HOME_JSON_LD } from './jsonLd';
 
 export default function Home() {
-    return <HomePage />;
+    return (
+        <>
+            <JsonLd name="home" data={HOME_JSON_LD} />
+            <HomePage />
+        </>
+    );
 }