+++ src/app/layout.tsx
@@ -1,8 +1,11 @@
+import type { PropsWithChildren } from 'react';
 import { AntdRegistry } from '@ant-design/nextjs-registry';
 import { ConfigProvider } from 'antd';
 import ruRu from 'antd/locale/ru_RU';
 import clsx from 'clsx';
+import { JsonLd } from 'features/seo/ui/JsonLD';
 import { body } from './font';
+import { ORGANIZATION_JSON_LD, WEB_SITE_JSON_LD } from './jsonLd';
 import { theme } from './theme';
 import { Modals } from './ui/Modals';
 import './styles/index.scss';
@@ -11,17 +14,18 @@ import styles from './styles.module.scss';
 
 export { metadata, viewport } from './meta';
 
-export default function RootLayout({ children }: { children: React.ReactNode }) {
+type Props = PropsWithChildren;
+
+export default function RootLayout(props: Props) {
+    const { children } = props;
+
     return (
         <html lang="ru">
-            {/* <head>
-                <link rel="icon" type="image/png" href="/favicon-96x96.png" sizes="96x96" />
-                <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
-                <link rel="shortcut icon" href="/favicon.ico" />
-                <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
-                <meta name="apple-mobile-web-app-title" content="Site" />
-                <link rel="manifest" href="/site.webmanifest" />
-            </head> */}
+            <head>
+                <meta name="apple-mobile-web-app-title" content="Title" />
+                <JsonLd name="organization" data={ORGANIZATION_JSON_LD} />
+                <JsonLd name="website" data={WEB_SITE_JSON_LD} />
+            </head>
             <body className={clsx(styles.body, body.className)}>
                 <AntdRegistry>
                     <ConfigProvider theme={theme} locale={ruRu}>