+++ src/features/seo/model/const/jsonLd.ts
@@ -1,0 +1,3 @@
+import { getAbsoluteUrl } from 'shared/lib/helpers/getAbsoluteUrl';
+
+export const ORGANIZATION_ID = getAbsoluteUrl('/#org');