+++ src/pages/example/CatalogPage/client.tsx
@@ -1,0 +1,10 @@
+'use client';
+
+import type { FC } from 'react';
+import { useCatalog } from 'entities/Example/api/swr/useCatalog';
+
+export const Client: FC = () => {
+    const { data = [] } = useCatalog();
+
+    return <>{data.map(({ name }) => name)}</>;
+};