+++ src/shared/types/page.ts
@@ -1,7 +1,7 @@
-export type IdParam = {
-    id: string;
-};
+export type IdParam = { id: string };
+export type SlugParam = { slug: string };
 
-export type PageId = {
-    params: IdParam;
-};
+export type PageParams<T> = { params: Promise<T> };
+
+export type PageId = PageParams<IdParam>;
+export type PageSlug = PageParams<SlugParam>;