+++ src/widgets/Header/index.tsx
@@ -1,6 +1,12 @@
 import type { FC } from 'react';
+import { DevLayout } from 'features/layout/DevLayout';
 import styles from './styles.module.scss';
 
 export const Header: FC = () => {
-    return <header className={styles.header}>Header</header>;
+    return (
+        <header className={styles.header}>
+            <DevLayout />
+            Header
+        </header>
+    );
 };